fix ingredient overlay
