+++ src/components/BurgerIngredients/BurgerIngredientsItem/BurgerIngredientsItem.jsx
@@ -5,21 +5,14 @@ import {
 import PropTypes from "prop-types";
 import React, { useCallback } from "react";
 import { useDrag } from "react-dnd";
-import { useDispatch, useSelector } from "react-redux";
+import { useDispatch } from "react-redux";
 import { useLocation, useNavigate } from "react-router-dom";
 import { dataPropTypes } from "../../../data/dataPropTypes";
-import { URL_INGREDIENTS, URL_ROOT } from "../../../data/routes";
-import naming from "../../../data/ru.json";
+import { URL_INGREDIENTS } from "../../../data/routes";
 import { INGREDIENTS_ACTIONS } from "../../../services/actions/ingredients-action";
-import Modal from "../../Modal/Modal";
 import styles from "./BurgerIngredientsItem.module.css";
-import IngredientDetails from "./IngredientDetails/IngredientDetails";
 
 const BurgerIngredientItem = ({ ingredient, count }) => {
-  const displayedIngredient = useSelector(
-    (state) => state.ingredientWindow.displayedIngredient
-  );
-
   const dispatch = useDispatch();
   const navigate = useNavigate();
   const location = useLocation();
@@ -36,15 +29,6 @@ const BurgerIngredientItem = ({ ingredient, count }) => {
     type: ingredient.type,
     item: ingredient,
   });
-
-  const hideIngredientDetails = useCallback(
-    (e) => {
-      navigate(URL_ROOT, { replace: true });
-      dispatch({ type: INGREDIENTS_ACTIONS.SHOW_DETAILS, item: null });
-      e.stopPropagation();
-    },
-    [dispatch, navigate]
-  );
 
   return (
     <li
@@ -68,14 +52,6 @@ const BurgerIngredientItem = ({ ingredient, count }) => {
       </div>
       {count > 0 && (
         <Counter count={count} size="default" extraClass={styles.count} />
-      )}
-      {displayedIngredient && (
-        <Modal
-          title={naming.IngredientDetails.title}
-          onClose={hideIngredientDetails}
-        >
-          <IngredientDetails item={displayedIngredient} />
-        </Modal>
       )}
     </li>
   );
